Add chunk enrichment mode and phase context to dashboard
Introduces chunk enrichment mode awareness and phase context to the dashboard UI, including new badges, banners, and progress bar logic. Backend API and stats endpoints now provide current worker phase and chunk enrichment mode, with improved ETA and status context. Also migrates image dimension columns to BIGINT in the database to support extreme values.

+++ backend/src/db/models.py
@@ -84,8 +84,8 @@ class RawFile(Base):
     vision_model = Column(Text)  # Which vision model was used
     
     # Image metadata
-    image_width = Column(Integer)
-    image_height = Column(Integer)
+    image_width = Column(BigInteger)
+    image_height = Column(BigInteger)
     
     # Series detection fields
     series_name = Column(Text)  # Detected series name
